+++ variables.tf
@@ -170,13 +170,6 @@ variable "restrict_public_buckets" {
   default     = true
 }
 
-
-# variable "website" {
-#   description = "Map containing static web-site hosting or redirect configuration."
-#   type        = any # map(string)
-#   default     = {}
-# }
-
 variable "lifecycle_rule" {
   description = "List of maps containing configuration of object lifecycle management."
   type        = any
@@ -249,12 +242,6 @@ variable "analytics_configuration" {
   default     = {}
 }
 
-# variable "cors_rule" {
-#   description = "List of maps containing rules for Cross-Origin Resource Sharing."
-#   type        = any
-#   default     = []
-# }
-
 variable "logging" {
   description = "Map containing access bucket logging configuration."
   type        = any
@@ -272,4 +259,3 @@ variable "acceleration_status" {
   type        = string
   default     = null
 }
-