+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # CHANGELOG - AWS S3 Module
 
-
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -12,7 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - An example for server-side configuration
 
 ### Changed
-- Remove support for website and CORS configuration
--
+- Remove support for website and CORS configuration examples